requirements.txt
# FastMCP 2.0 智能工具调度系统 - 依赖包

# 核心框架
fastmcp>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# 数据库
pymysql>=1.1.0
sqlalchemy>=2.0.0

# 文档处理
python-docx>=1.1.0

# 向量数据库
chromadb>=0.4.18

# 知识图谱
networkx>=3.1

# OpenAI客户端
openai>=1.3.0

# 数据处理
pydantic>=2.5.0
python-multipart>=0.0.6

# HTTP请求
requests>=2.31.0
httpx>=0.25.0

# 时间处理
pytz>=2023.3
python-dateutil>=2.8.2

# 日志和配置
python-dotenv>=1.0.0

# 其他工具
aiofiles>=23.2.0
rich>=13.7.0
